Relocate MCP tool definitions and loader to `lib/mcp` for improved modularity and organization
