+++ forge/ee/lib/mcp/toolLoader.js
